Revert "Construct NJT without graph breaks"
This reverts commit 911154271309667b55dfb963ec6384bd0048019b.

[ghstack-poisoned]
