+++ torch/_subclasses/fake_tensor.py
@@ -531,13 +531,8 @@ class FakeTensorConfig:
 # Making this a descriptor may seem overly fancy, but actually it's the most
 # convenient way to make sure we have access to FakeTensor during access,
 # which is required for testing version counter and epoch validity
-class SymIntMemoDescriptor:
+class UnbackedMemoDescriptor:
     _name: str
-    _is_unbacked: bool
-
-    def __init__(self, *, is_unbacked: Optional[bool] = None):
-        assert is_unbacked is not None
-        self._is_unbacked = is_unbacked
 
     def __set_name__(self, owner: str, name: str) -> None:
         self._name = name
@@ -557,20 +552,20 @@ class SymIntMemoDescriptor:
 
     def __get__(
         self, obj: FakeTensor, objtype: Optional[Type[FakeTensor]] = None
-    ) -> Optional[torch.SymInt]:
+    ) -> Optional[object]:
         if (r := getattr(obj, self._memo(obj))) is None:
             return None
         # Version counter based tracking isn't 100% sound but it's close
         # enough
-        if getattr(obj, self._memo_vc(obj)) != obj._version or (
-            self._is_unbacked
-            and getattr(obj, self._memo_epoch(obj)) != obj.fake_mode.epoch
+        if (
+            getattr(obj, self._memo_vc(obj)) != obj._version
+            or getattr(obj, self._memo_epoch(obj)) != obj.fake_mode.epoch
         ):
             setattr(obj, self._memo(obj), None)
             return None
         return r
 
-    def __set__(self, obj: FakeTensor, value: Optional[torch.SymInt]) -> None:
+    def __set__(self, obj: FakeTensor, value: Optional[object]) -> None:
         if value is None:
             setattr(obj, self._memo(obj), None)
             setattr(obj, self._memo_vc(obj), None)
@@ -598,14 +593,9 @@ class FakeTensor(Tensor):
     # TODO: Generalize this as needed, e.g., into a trie of memos, if
     # you do something like x[0].item()  (x[0] is fresh each time, so
     # memo mechanism here won't work)
-    nonzero_memo = SymIntMemoDescriptor(is_unbacked=True)
-    item_memo = SymIntMemoDescriptor(is_unbacked=True)
-    unique_memo = SymIntMemoDescriptor(is_unbacked=True)
-
-    # We expect nested_int_memo to be None when an offsets is a graph
-    # intermediate, or an input that has never been associated with a
-    # nested int.
-    nested_int_memo = SymIntMemoDescriptor(is_unbacked=False)
+    nonzero_memo = UnbackedMemoDescriptor()
+    item_memo = UnbackedMemoDescriptor()
+    unique_memo = UnbackedMemoDescriptor()
 
     # Indicates to our torch_dispatch dispatching infra that
     # this is an "infra" mode with lower dispatching precedence.
@@ -703,7 +693,6 @@ class FakeTensor(Tensor):
         self.nonzero_memo = None
         self.item_memo = None
         self.unique_memo = None
-        self.nested_int_memo = None
 
         if FakeTensorConfig.debug:
             self._debug_trace = CapturedTraceback.extract()  # type: ignore[attr-defined]
@@ -873,17 +862,6 @@ class FakeTensor(Tensor):
         assert common_device is not None, f"Could not find common device for {func}"
 
         return common_device, has_scalar_only_inputs
-
-    def get_nested_int(
-        self,
-        *,
-        coeff: Union[int, torch.SymInt] = 1,
-    ) -> torch.SymInt:
-        if self.nested_int_memo is None:
-            self.nested_int_memo = self.fake_mode.create_symbolic_nested_int(
-                nt_tensor_id=None
-            )
-        return self.nested_int_memo * coeff
 
     # We must handle tolist in a special way for FakeTensors here in the case
     # where tolist is called from torch dispatch for tensor subclasses.
@@ -1085,16 +1063,6 @@ class FakeTensorMode(TorchDispatchMode):
     _stack: Optional[str]
     allow_meta: bool
 
-    # NestedTensor uses a tensor_id_counter to uniquely identify offsets.
-    # This counter is incremented when an offsets is used to create an NJT
-    # for the first time. To avoid mutating eager state if we construct NJT
-    # during tracing, we maintain a separate counter on the FakeTensorMode.
-    # The initial count is set to the current eager tensor_id_counter value
-    # upon initialization, and every time you retrace using the same fake tensor
-    # mode, you should reset the counter to the initial count.
-    nt_tensor_id_counter: int = -1
-    nt_tensor_id_initial_count: int = -1
-
     def __init__(
         self,
         *,
@@ -1183,16 +1151,6 @@ class FakeTensorMode(TorchDispatchMode):
         # this is an "infra" mode with lower dispatching precedence.
         self._mode_key = torch._C._TorchDispatchModeKey.FAKE
 
-        import torch.nested._internal.nested_tensor
-
-        self.nt_tensor_id_initial_count = (
-            torch.nested._internal.nested_tensor._tensor_id_counter
-        )
-        self.nt_tensor_id_counter = self.nt_tensor_id_initial_count
-
-    def reset_nt_tensor_id_counter(self) -> None:
-        self.nt_tensor_id_counter = self.nt_tensor_id_initial_count
-
     # Typically, there is only one fake tensor mode and you test for it by
     # doing an isinstance test.  However, in some situations, there might be
     # TWO fake tensor modes.  The canonical example of this is exporting
@@ -1247,8 +1205,6 @@ class FakeTensorMode(TorchDispatchMode):
 
     # No-op if FakeTensorMode is already in use
     def __enter__(self) -> Self:
-        import torch.nested._internal.nested_tensor
-
         prev_only_lift_cpu_tensors = None
         if self.avoid_device_init:
             # See NOTE: [torch.tensor, lift_fresh, and device movement]
@@ -2143,31 +2099,6 @@ class FakeTensorMode(TorchDispatchMode):
                 return e
 
         return tree_map(wrap, r)
-
-    def create_symbolic_nested_int(
-        self, *, nt_tensor_id: Optional[int] = None
-    ) -> torch.SymInt:
-        # See Note: [Creating symbolic nested int]
-        # Returned nested int always has coeff=1; multiply the result by coeff if needed
-        import torch.nested._internal.nested_tensor
-
-        if nt_tensor_id is None:
-            nt_tensor_id = self.nt_tensor_id_counter
-            assert self.enter_stack, "should only called while FakeTensorMode is active"
-            self.nt_tensor_id_counter += 1
-        hint = torch._C._get_nested_int(nt_tensor_id, 1)
-
-        src = torch._dynamo.source.EphemeralSource("intermediate_offsets_or_lengths")
-        assert self.shape_env is not None
-        ret = self.shape_env.create_symintnode(
-            sym=self.shape_env.create_symbol(
-                val=hint,
-                source=src,
-            ),
-            hint=hint,
-            source=src,
-        )
-        return ret
 
     _cpp_meta_supports_symint = ordered_set(
         aten.empty.memory_format,
