+++ torch/nested/_internal/ops.py
@@ -511,20 +511,7 @@ def _to_copy_default(func, *args, **kwargs):
 
     # Copy to a new Python subclass NestedTensor
     new_offsets = inp._offsets.to(device=new_values.device)
-
-    from torch._subclasses.fake_tensor import FakeTensor
-    from torch._subclasses.functional_tensor import (
-        FunctionalTensor,
-        mb_unwrap_functional_tensor,
-    )
-
-    if isinstance(new_offsets, (FakeTensor, FunctionalTensor)):
-        # Temporary hack until we have the union find
-        tgt = mb_unwrap_functional_tensor(new_offsets)
-        src = mb_unwrap_functional_tensor(inp._offsets)
-        tgt.nested_int_memo = src.nested_int_memo
-    else:
-        _tensor_symint_registry[new_offsets] = _tensor_symint_registry[inp._offsets]
+    _tensor_symint_registry[new_offsets] = _tensor_symint_registry[inp._offsets]
     inp_kwargs = extract_kwargs(inp)
     inp_kwargs["offsets"] = new_offsets
